+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - SubScopeManager can dynamically create new sub scopes
 - Node and scope names must be camel case now
 - Scope:instantiate: Rename parentScope param into scope
+- ScopeBluePrints will alwas create a scope
 
 ### Removed